+++ jest.config.js
@@ -8,6 +8,10 @@ module.exports = {
         "<rootDir>/src"
     ],
 
+    roots: [
+        "<rootDir>/src"
+    ],
+
     collectCoverageFrom: [
         "src/**/*.js"
     ],
@@ -18,7 +22,7 @@ module.exports = {
         "/node_modules/"
     ],
 
-    setupTestFrameworkScriptFile: "<rootDir>/jest.setup.js",
+    setupFilesAfterEnv: ["<rootDir>/jest.setup.js"],
 
     snapshotSerializers: ["enzyme-to-json/serializer"],