+++ src/DensityChart/PlayButton.js
@@ -2,15 +2,15 @@ import React, { PureComponent } from "react";
 import PropTypes from "prop-types";
 
 // These icon paths were copied from antd
-const PAUSE_ICON_PATH = `M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 
-64 512 64zm0 820c-205.4 0-372-166.6-372-372s166.6-372 372-372 372 166.6 372 372-166.6 372-372 
-372zm-88-532h-48c-4.4 0-8 3.6-8 8v304c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8V360c0-4.4-3.6-8-8-8zm224 
+const PAUSE_ICON_PATH = `M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4
+64 512 64zm0 820c-205.4 0-372-166.6-372-372s166.6-372 372-372 372 166.6 372 372-166.6 372-372
+372zm-88-532h-48c-4.4 0-8 3.6-8 8v304c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8V360c0-4.4-3.6-8-8-8zm224
 0h-48c-4.4 0-8 3.6-8 8v304c0 4.4 3.6 8 8 8h48c4.4 0 8-3.6 8-8V360c0-4.4-3.6-8-8-8z`;
 
 const PLAY_ICON_PATHS = [
-    `M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 64 512 64zm0 820c-205.4 
+    `M512 64C264.6 64 64 264.6 64 512s200.6 448 448 448 448-200.6 448-448S759.4 64 512 64zm0 820c-205.4
     0-372-166.6-372-372s166.6-372 372-372 372 166.6 372 372-166.6 372-372 372z`,
-    `M719.4 499.1l-296.1-215A15.9 15.9 0 0 0 398 297v430c0 13.1 14.8 20.5 25.3 12.9l296.1-215a15.9 
+    `M719.4 499.1l-296.1-215A15.9 15.9 0 0 0 398 297v430c0 13.1 14.8 20.5 25.3 12.9l296.1-215a15.9
     15.9 0 0 0 0-25.8zm-257.6 134V390.9L628.5 512 461.8 633.1z`
 ];
 
@@ -26,7 +26,6 @@ const PLAY_ICON_PATHS = [
  */
 
 export default class PlayButton extends PureComponent {
-
     static propTypes = {
         width: PropTypes.number.isRequired,
         brushDomainMax: PropTypes.number.isRequired,
@@ -151,14 +150,14 @@ export default class PlayButton extends PureComponent {
         if (this.state.play) {
             iconElement = (
                 <svg viewBox="64 64 896 896" width="1em" height="1em" fill="currentColor">
-                    <path d={PAUSE_ICON_PATH}></path>
+                    <path d={PAUSE_ICON_PATH}/>
                 </svg>
             );
         } else {
             iconElement = (
                 <svg viewBox="64 64 896 896" width="1em" height="1em" fill="currentColor">
-                    <path d={PLAY_ICON_PATHS[0]}></path>
-                    <path d={PLAY_ICON_PATHS[1]}></path>
+                    <path d={PLAY_ICON_PATHS[0]}/>
+                    <path d={PLAY_ICON_PATHS[1]}/>
                 </svg>
             );
         }